fix(tests): directory
