+++ .github/workflows/production.yml
@@ -6,7 +6,7 @@ jobs:
   go-lint:
     strategy:
       matrix:
-        go-version: [1.16.x]
+        go-version: [1.19.x]
         platform: [ubuntu-latest]
     runs-on: ${{ matrix.platform }}
     steps:
@@ -34,7 +34,7 @@ jobs:
     needs: [ go-lint ]
     strategy:
       matrix:
-        go-version: [1.17.x,1.18.x,1.19.x]
+        go-version: [1.18.x,1.19.x]
         platform: [ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.platform }}
     steps:
@@ -57,7 +57,7 @@ jobs:
         run: go get -u 
 
       - name: Test
-        run: go test -v
+        run: go test -v ./... 
 
   build-docker-artifacts:
     needs: [ unit-test ]
@@ -66,7 +66,7 @@ jobs:
     steps:
       - uses: actions/setup-go@v1
         with:
-          go-version: '1.15.x'
+          go-version: '1.19.x'
 
       - uses: actions/checkout@v1
